+++ gazebo/physics/WorldState.hh
@@ -70,6 +70,14 @@ namespace gazebo
       /// \param[in] _world Pointer to a world
       public: void Load(const WorldPtr _world);
 
+      /// \brief Load from a World pointer.
+      ///
+      /// Generate a WorldState from an instance of a World.
+      /// \param[in] _world Pointer to a world
+      /// \param[in] _filter String for filtering models states
+      public: void LoadWithFilter(const WorldPtr _world,
+          const std::string &_filter);
+
       /// \brief Load state from SDF element.
       ///
       /// Set a WorldState from an SDF element containing WorldState info.