+++ gazebo/physics/dart/DARTModel.cc
@@ -293,26 +293,6 @@ void DARTModel::Fini()
   {
     _world->removeSkeleton(this->dataPtr->dtSkeleton);
   }
-
-  // update the world which now does not have the removed model an more
-  if (_world)
-  {
-    // We need to clear the last collision result because the next call of
-    // DARTPhysics::UpdateCollision() crashes if the last collision result
-    // still contains the removed model.
-    // Unfortunately, we cannot use
-    // _world->getLastCollisionResult().clear()
-    // (DARTPhysics::UpdateCollision() uses world->getLastCollisionResult())
-    // because it returns a const pointer.
-    // We could instead use
-    // _world->getConstraintSolver()->getLastCollisionResult().clear();
-    // (which works as well, as tested with DART-6)
-    // but that could conflict with future versions of DART and different
-    // handling in DARTPhysics::UpdateCollision() and here.
-    // So instead, for now we will just step() the world, which will also
-    // update the last collision result used in DARTPhysics::UpdateCollision().
-    _world->step();
-  }
 }
 
 //////////////////////////////////////////////////