+++ gazebo/util/LogRecord_TEST.cc
@@ -185,6 +185,32 @@ TEST_F(LogRecord_TEST, Start_zlib)
 }
 
 /////////////////////////////////////////////////
+/// \brief Test LogRecord filter
+TEST_F(LogRecord_TEST, Filter)
+{
+  gazebo::util::LogRecord *recorder = gazebo::util::LogRecord::Instance();
+
+  // check default values
+  EXPECT_DOUBLE_EQ(recorder->Period(), -1);
+  EXPECT_EQ(recorder->Filter(), std::string());
+
+  // filter by period
+  recorder->SetPeriod(0.02);
+  EXPECT_DOUBLE_EQ(recorder->Period(), 0.02);
+
+  recorder->SetPeriod(0);
+  EXPECT_DOUBLE_EQ(recorder->Period(), 0);
+
+
+  // filter by regex string
+  recorder->SetFilter("robot*");
+  EXPECT_EQ(recorder->Filter(), "robot*");
+
+  recorder->SetFilter("");
+  EXPECT_EQ(recorder->Filter(), "");
+}
+
+/////////////////////////////////////////////////
 int main(int argc, char **argv)
 {
   ::testing::InitGoogleTest(&argc, argv);