+++ test/integration/physics_collision.cc
@@ -43,8 +43,7 @@ class PhysicsCollisionTest : public ServerFixture,
 /////////////////////////////////////////////////
 void PhysicsCollisionTest::GetBoundingBox(const std::string &_physicsEngine)
 {
-  if (_physicsEngine == "simbody" ||
-      _physicsEngine == "dart")
+  if (_physicsEngine == "simbody")
   {
     gzerr << "Bounding boxes not yet working with "
           << _physicsEngine
@@ -57,16 +56,72 @@ void PhysicsCollisionTest::GetBoundingBox(const std::string &_physicsEngine)
   physics::WorldPtr world = physics::get_world("default");
   ASSERT_TRUE(world != NULL);
 
-  // Check bounding box of ground plane
+  // Check bounding box of ground plane.
+  // Skip this test for DART because planes aren't handled properly yet.
+  // For this test to pass for DART, the following values would need to be set:
+  //    big = 1049;
+  //    z = 1050;
+  if (_physicsEngine != "dart")
   {
+    double big = g_big;
+    double z = 0.0;
+
     physics::ModelPtr model = world->ModelByName("ground_plane");
+    ASSERT_TRUE(model != nullptr);
     ignition::math::Box box = model->BoundingBox();
-    EXPECT_LT(box.Min().X(), -g_big);
-    EXPECT_LT(box.Min().Y(), -g_big);
-    EXPECT_LT(box.Min().Z(), -g_big);
-    EXPECT_GT(box.Max().X(), g_big);
-    EXPECT_GT(box.Max().Y(), g_big);
-    EXPECT_DOUBLE_EQ(box.Max().Z(), 0.0);
+    EXPECT_LT(box.Min().X(), -big);
+    EXPECT_LT(box.Min().Y(), -big);
+    EXPECT_LT(box.Min().Z(), -big);
+    EXPECT_GT(box.Max().X(), big);
+    EXPECT_GT(box.Max().Y(), big);
+    EXPECT_DOUBLE_EQ(box.Max().Z(), z);
+  }
+
+  // Insert model which forms a t-shape, made up of two long boxes.
+  // The second box will have to be translated relative to the first to
+  // form the "roof" of the T. This tests whether the relative transforms
+  // of the collision shapes are considered when computing the bounding box.
+  // If they are not, the result would be that of a cross-shape instead.
+
+  std::ostringstream sdfStream;
+  sdfStream << "<sdf version='" << SDF_VERSION << "'>"
+    << "<model name ='model_tshape'>"
+    << "<static> true </static>"
+    << "<pose>0 0 0 0 0 0</pose>"
+    << "<link name=\"link\">"
+      << "<collision name=\"pose\">"
+        << "<pose>0 0.5 0 0 0 0</pose>"
+        << "<geometry>"
+          << "<box>"
+            << "<size>0.1 1 0.1</size>"
+          << "</box>"
+        << "</geometry>"
+      << "</collision>"
+      << "<collision name=\"top\">"
+        << "<pose>0 0.95 0 0 0 0</pose>"
+        << "<geometry>"
+          << "<box>"
+            << "<size>1 0.1 0.1</size>"
+          << "</box>"
+        << "</geometry>"
+      << "</collision>"
+    << "</link>"
+    << "</model>"
+    << "</sdf>";
+  SpawnSDF(sdfStream.str());
+
+  {
+    physics::ModelPtr model = world->ModelByName("model_tshape");
+    ASSERT_TRUE(model != nullptr);
+    ignition::math::Box box = model->BoundingBox();
+    gzdbg << "Bounding box for " << _physicsEngine << ": " << box << std::endl;
+    static double tol = 1e-03;
+    EXPECT_NEAR(box.Min().X(), -0.5, tol);
+    EXPECT_NEAR(box.Min().Y(), 0, tol);
+    EXPECT_NEAR(box.Min().Z(), -0.05, tol);
+    EXPECT_NEAR(box.Max().X(), 0.5, tol);
+    EXPECT_NEAR(box.Max().Y(), 1, tol);
+    EXPECT_NEAR(box.Max().Z(), 0.05, tol);
   }
 }
 
@@ -102,6 +157,20 @@ void PhysicsCollisionTest::PoseOffsets(const std::string &_physicsEngine)
     auto collision = link->GetCollision(index);
     ASSERT_TRUE(collision != nullptr);
 
+    // for box_3, also get link_2 and its collision
+    auto link2 = model->GetLink("link_2");
+    physics::CollisionPtr collision2;
+    if (i == 3)
+    {
+      EXPECT_NE(link2, nullptr);
+      collision2 = link2->GetCollision(index);
+      EXPECT_NE(collision2, nullptr);
+    }
+    else
+    {
+      EXPECT_EQ(link2, nullptr);
+    }
+
     // i=0: rotated model pose
     //  expect collision pose to match model pose
     if (i == 0)
@@ -127,6 +196,10 @@ void PhysicsCollisionTest::PoseOffsets(const std::string &_physicsEngine)
       ignition::math::Pose3d collisionPose(0, 0, dz, 0, 0, 0);
       EXPECT_EQ(link->WorldPose().Pos() + collisionPose.Pos(),
                 collision->WorldPose().Pos());
+
+      // collision2 pose should match link2 pose
+      EXPECT_EQ(link2->WorldPose().Pos(),
+                collision2->WorldPose().Pos());
     }
   }
   EXPECT_EQ(modelCount, 4);
@@ -154,6 +227,20 @@ void PhysicsCollisionTest::PoseOffsets(const std::string &_physicsEngine)
     auto collision = link->GetCollision(index);
     ASSERT_TRUE(collision != nullptr);
 
+    // for box_3, also get link_2 and its collision
+    auto link2 = model->GetLink("link_2");
+    physics::CollisionPtr collision2;
+    if (i == 3)
+    {
+      EXPECT_NE(link2, nullptr);
+      collision2 = link2->GetCollision(index);
+      EXPECT_NE(collision2, nullptr);
+    }
+    else
+    {
+      EXPECT_EQ(link2, nullptr);
+    }
+
     // For 0-2, drop and expect box to rest at specific height
     if (i <= 2)
     {
@@ -162,6 +249,7 @@ void PhysicsCollisionTest::PoseOffsets(const std::string &_physicsEngine)
     else
     {
       EXPECT_NEAR(collision->WorldPose().Pos().Z(), dz/2, g_physics_tol);
+      EXPECT_NEAR(collision2->WorldPose().Pos().Z(), dz/2, g_physics_tol);
     }
   }
 }