+++ test/integration/physics_link.cc
@@ -24,12 +24,14 @@
 #include "gazebo/test/ServerFixture.hh"
 #include "gazebo/test/helper_physics_generator.hh"
 
+#include "test/util.hh"
+
 using namespace gazebo;
 
 const double g_tolerance = 1e-4;
 
 class PhysicsLinkTest : public ServerFixture,
-                        public testing::WithParamInterface<const char*>
+                        public ::testing::WithParamInterface<const char*>
 {
   /// \brief Test force adding functions.
   /// \param[in] _physicsEngine Type of physics engine to use.
@@ -37,12 +39,19 @@ class PhysicsLinkTest : public ServerFixture,
 
   /// \brief Use AddLinkForce on the given direction and then the opposite
   /// direction so they cancel out.
+  /// \param[in] _physicsEngine Name of the physics engine that is being used
+  /// \param[in] _world_equals_link World Frame == Link Frame
+  /// \param[in] _link_equals_inertial Link Frame == Inertial Frame
   /// \param[in] _world World pointer.
   /// \param[in] _link Link pointer.
   /// \param[in] _force Force expressed in link frame.
   /// \param[in] _offset Offset expressed in link frame, defaults to link
   /// origin.
-  public: void AddLinkForceTwoWays(physics::WorldPtr _world,
+  public: void AddLinkForceTwoWays(
+      const std::string &_physicsEngine,
+      const bool _world_equals_link,
+      const bool _link_equals_inertial,
+      physics::WorldPtr _world,
       physics::LinkPtr _link, ignition::math::Vector3d _force,
       ignition::math::Vector3d _offset = ignition::math::Vector3d::Zero);
 
@@ -68,7 +77,11 @@ class PhysicsLinkTest : public ServerFixture,
 };
 
 /////////////////////////////////////////////////
-void PhysicsLinkTest::AddLinkForceTwoWays(physics::WorldPtr _world,
+void PhysicsLinkTest::AddLinkForceTwoWays(
+    const std::string& _physicsEngine,
+    const bool _world_equals_link,
+    const bool _link_equals_inertial,
+    physics::WorldPtr _world,
     physics::LinkPtr _link, ignition::math::Vector3d _force,
     ignition::math::Vector3d _offset)
 {
@@ -83,12 +96,27 @@ void PhysicsLinkTest::AddLinkForceTwoWays(physics::WorldPtr _world,
   else
     _link->AddLinkForce(_force, _offset);
 
-  double dt = _world->Physics()->GetMaxStepSize();
-  _world->Step(1);
+  const int moreThanOneStep = 2;
+  const double dt = _world->Physics()->GetMaxStepSize();
 
-  int moreThanOneStep = 2;
+  // Note: This step must be performed before checking the link forces when ODE
+  // is the physics engine, because otherwise the link accelerations will not be
+  // computed, and therefore the result of WorldForce and WorldTorque (which
+  // depend on the current accelerations) will be zero.
+  //
+  // However, other simulation engines (such as DART) maintain the link forces
+  // and torques as part of the current state, and the values (might) get
+  // cleared out after each simulation step (depending on the engine's current
+  // settings), so it should not necessarily be called for each engine. If this
+  // test is enabled for another engine (such as SimBody or Bullet), then that
+  // engine should perform its simulation step either in this if-statement or
+  // the one below depending on how its simulation procedure works.
+  if ("ode" == _physicsEngine)
+  {
+    _world->Step(1);
+  }
 
-  // Check force and torque (at CoG?) in world frame
+  // Check force and torque relative to the COG in world coordinates
   ignition::math::Vector3d forceWorld = poseWorld0.Rot().RotateVector(_force);
   EXPECT_EQ(forceWorld, _link->WorldForce());
 
@@ -110,10 +138,34 @@ void PhysicsLinkTest::AddLinkForceTwoWays(physics::WorldPtr _world,
       _link->WorldInertiaMatrix().Inverse() * torqueWorld;
   EXPECT_EQ(oneStepAngularAccel, _link->WorldAngularAccel());
 
+
+  // Note: This step must be performed after checking the link forces when DART
+  // is the physics engine, because otherwise the accelerations used by the
+  // previous tests will be cleared out before they can be tested.
+  if ("dart" == _physicsEngine)
+  {
+    _world->Step(1);
+  }
+
   // Check velocity in world frame
   ignition::math::Vector3d oneStepLinearVel = linearVelWorld0 +
     dt*oneStepLinearAccel;
-  EXPECT_EQ(oneStepLinearVel, _link->WorldCoGLinearVel());
+
+  // Dev note (MXG): DART does not always produce quite the same result as the
+  // expected value for CoG linear velocity. It might be worth investigating
+  // whether this is ordinary numerical error or if DART should be tweaked to be
+  // more precise.
+  //
+  // The tests succeed for a tolerance as low as 1e-6 in all trials except
+  // "World != link != inertial frame". This makes me suspect that the
+  // inaccuracy is a result of cummulative matrix multiplications.
+  double tolerance = 1e-6;
+  if ("dart" == _physicsEngine && !_world_equals_link && !_link_equals_inertial)
+  {
+    tolerance = 2e-3;
+  }
+
+  VEC_EXPECT_NEAR(oneStepLinearVel, _link->WorldCoGLinearVel(), tolerance);
 
   ignition::math::Vector3d oneStepAngularVel = angularVelWorld0 +
     dt*oneStepAngularAccel;
@@ -131,7 +183,9 @@ void PhysicsLinkTest::AddLinkForceTwoWays(physics::WorldPtr _world,
   EXPECT_EQ(ignition::math::Vector3d::Zero, _link->WorldAngularAccel());
 
   // Check that velocity hasn't changed
-  EXPECT_EQ(oneStepLinearVel, _link->WorldCoGLinearVel());
+  // Dev note (MXG): Same note as above regarding the tolerance that is used
+  // here. Everything but "World != link != inertial frame" succeeds with 1e-6.
+  VEC_EXPECT_NEAR(oneStepLinearVel, _link->WorldCoGLinearVel(), tolerance);
   EXPECT_EQ(oneStepAngularVel, _link->WorldAngularVel());
 
   // Add opposing force in link frame and check that link is back to initial
@@ -153,11 +207,11 @@ void PhysicsLinkTest::AddLinkForceTwoWays(physics::WorldPtr _world,
 /////////////////////////////////////////////////
 void PhysicsLinkTest::AddForce(const std::string &_physicsEngine)
 {
-  // TODO bullet, dart and simbody currently fail this test
-  if (_physicsEngine != "ode")
+  // TODO bullet and simbody currently fail this test
+  if (_physicsEngine != "ode" && _physicsEngine != "dart")
   {
-    gzerr << "Aborting AddForce test for Bullet, DART and Simbody. "
-          << "See issues #1476, #1477, and #1478."
+    gzerr << "Aborting AddForce test for Bullet and Simbody. "
+          << "See issues #1476 and #1478."
           << std::endl;
     return;
   }
@@ -195,7 +249,8 @@ void PhysicsLinkTest::AddForce(const std::string &_physicsEngine)
   gzdbg << "World == link == inertial frames, no offset" << std::endl;
   EXPECT_EQ(ignition::math::Pose3d::Zero, link->WorldPose());
   EXPECT_EQ(ignition::math::Pose3d::Zero, link->WorldInertialPose());
-  this->AddLinkForceTwoWays(world, link, ignition::math::Vector3d(1, 20, 31));
+  this->AddLinkForceTwoWays(_physicsEngine, true, true, world, link,
+                            ignition::math::Vector3d(1, 20, 31));
 
   gzdbg << "World != link == inertial frames, no offset" << std::endl;
   model->SetLinkWorldPose(ignition::math::Pose3d(
@@ -203,14 +258,16 @@ void PhysicsLinkTest::AddForce(const std::string &_physicsEngine)
         ignition::math::Quaterniond(0, IGN_PI/2.0, 1)), link);
   EXPECT_NE(ignition::math::Pose3d::Zero, link->WorldPose());
   EXPECT_EQ(link->WorldPose(), link->WorldInertialPose());
-  this->AddLinkForceTwoWays(world, link, ignition::math::Vector3d(-1, 10, 5));
+  this->AddLinkForceTwoWays(_physicsEngine, false, true, world, link,
+                            ignition::math::Vector3d(-1, 10, 5));
 
   gzdbg << "World == link == inertial frames, with offset" << std::endl;
   model->SetLinkWorldPose(ignition::math::Pose3d::Zero, link);
   EXPECT_EQ(ignition::math::Pose3d::Zero, link->WorldPose());
   EXPECT_EQ(ignition::math::Pose3d::Zero, link->WorldInertialPose());
-  this->AddLinkForceTwoWays(world, link, ignition::math::Vector3d(5, 4, 3),
-      ignition::math::Vector3d(-2, 1, 0));
+  this->AddLinkForceTwoWays(_physicsEngine, true, true, world, link,
+                            ignition::math::Vector3d(5, 4, 3),
+                            ignition::math::Vector3d(-2, 1, 0));
 
   gzdbg << "World == link != inertial frames, no offset" << std::endl;
   model->SetLinkWorldPose(ignition::math::Pose3d::Zero, link);
@@ -218,9 +275,11 @@ void PhysicsLinkTest::AddForce(const std::string &_physicsEngine)
       ignition::math::Vector3d(1, 5, 8),
       ignition::math::Quaterniond(IGN_PI/3.0, IGN_PI*1.5, IGN_PI/4));
   link->GetInertial()->SetCoG(inertialPose);
+  link->UpdateMass();
   EXPECT_EQ(ignition::math::Pose3d::Zero, link->WorldPose());
   EXPECT_EQ(inertialPose, link->WorldInertialPose());
-  this->AddLinkForceTwoWays(world, link, ignition::math::Vector3d(1, 2, 1));
+  this->AddLinkForceTwoWays(_physicsEngine, true, false, world, link,
+                            ignition::math::Vector3d(1, 2, 1));
 
   gzdbg << "World != link != inertial frames, with offset" << std::endl;
   model->SetLinkWorldPose(ignition::math::Pose3d(
@@ -229,8 +288,10 @@ void PhysicsLinkTest::AddForce(const std::string &_physicsEngine)
   inertialPose = ignition::math::Pose3d(ignition::math::Vector3d(0, -5, 10),
       ignition::math::Quaterniond(0, 2.0*IGN_PI, IGN_PI/3));
   link->GetInertial()->SetCoG(inertialPose);
-  this->AddLinkForceTwoWays(world, link, ignition::math::Vector3d(1, 2, 1),
-      ignition::math::Vector3d(-2, 0.5, 1));
+  link->UpdateMass();
+  this->AddLinkForceTwoWays(_physicsEngine, false, false, world, link,
+                            ignition::math::Vector3d(1, 2, 1),
+                            ignition::math::Vector3d(-2, 0.5, 1));
 
   gzdbg << "World != link != inertial frames, with offset and initial vel"
       << std::endl;
@@ -241,10 +302,12 @@ void PhysicsLinkTest::AddForce(const std::string &_physicsEngine)
   inertialPose = ignition::math::Pose3d(ignition::math::Vector3d(1, 0, -5.6),
       ignition::math::Quaterniond(IGN_PI/9, 0, IGN_PI*3));
   link->GetInertial()->SetCoG(inertialPose);
+  link->UpdateMass();
   link->SetLinearVel(ignition::math::Vector3d(2, -0.1, 5));
   link->SetAngularVel(ignition::math::Vector3d(-IGN_PI/10, 0, 0.0001));
-  this->AddLinkForceTwoWays(world, link, ignition::math::Vector3d(-3, 2.5, -15),
-      ignition::math::Vector3d(-6, -1, -0.2));
+  this->AddLinkForceTwoWays(_physicsEngine, false, false, world, link,
+                            ignition::math::Vector3d(-3, 2.5, -15),
+                            ignition::math::Vector3d(-6, -1, -0.2));
 }
 
 /////////////////////////////////////////////////
@@ -505,17 +568,11 @@ void PhysicsLinkTest::GetWorldInertia(const std::string &_physicsEngine)
         link->SetTorque(ignition::math::Vector3d(100, 0, 0));
         world->Step(1);
       }
-      if (_physicsEngine.compare("dart") == 0)
-      {
-        gzerr << "Dart fails this portion of the test (#1090)" << std::endl;
-      }
-      else
-      {
-        ignition::math::Vector3d vel = link->WorldAngularVel();
-        EXPECT_NEAR(vel.X(),  0.0703, g_tolerance);
-        EXPECT_NEAR(vel.Y(), -0.0049, g_tolerance);
-        EXPECT_NEAR(vel.Z(),  0.0000, g_tolerance);
-      }
+
+      ignition::math::Vector3d vel = link->WorldAngularVel();
+      EXPECT_NEAR(vel.X(),  0.0703, g_tolerance);
+      EXPECT_NEAR(vel.Y(), -0.0049, g_tolerance);
+      EXPECT_NEAR(vel.Z(),  0.0000, g_tolerance);
     }
   }
 }
@@ -523,12 +580,11 @@ void PhysicsLinkTest::GetWorldInertia(const std::string &_physicsEngine)
 /////////////////////////////////////////////////
 void PhysicsLinkTest::OnWrenchMsg(const std::string &_physicsEngine)
 {
-  // TODO bullet, dart and simbody currently fail this test
-  if (_physicsEngine != "ode")
+  // TODO bullet and simbody currently fail this test
+  if (_physicsEngine != "ode" && _physicsEngine != "dart")
   {
-    gzerr << "Aborting OnWrenchMsg test for Bullet, DART and Simbody. "
-          << "Because of issues #1476, #1477, and #1478."
-          << std::endl;
+    gzerr << "Aborting OnWrenchMsg test for Bullet and Simbody. "
+          << "Because of issues #1476 and #1478." << std::endl;
     return;
   }
 
@@ -595,6 +651,18 @@ void PhysicsLinkTest::OnWrenchMsg(const std::string &_physicsEngine)
   torques.push_back(ignition::math::Vector3d(-0.2, 5, 0));
   forceOffsets.push_back(ignition::math::Vector3d(-1, -4, -0.8));
 
+  // When DART is the phyiscs engine, in order to pass this test, we need to
+  // turn off the default behavior of clearing the forces and torques after
+  // each simulation step. The force and torques that are sent over messages
+  // don't get received by the simulation engine until the update step occurs,
+  // but the update step will also simulate forward, and DART's default behavior
+  // would then clear out the forces and torques. Setting this parameter
+  // overrides that behavior.
+  if ("dart" == _physicsEngine)
+  {
+    physics->SetParam("auto_reset_forces", false);
+  }
+
   for (unsigned int i = 0; i < forces.size(); ++i)
   {
     gzdbg << "Testing force: " << forces[i].X() << ", "
@@ -623,17 +691,28 @@ void PhysicsLinkTest::OnWrenchMsg(const std::string &_physicsEngine)
     ignition::math::Vector3d torqueWorld =
       worldOffset.Cross(forces[i]) + torques[i];
 
+    double tolerance = 1e-6;
+    // Note: Similar to the issue in AddLinkForceTwoWays, DART experiences some
+    // inaccuracies when returning the WorldForce. I suspect this is because of
+    // numerical inaccuracies built up through matrix multiplications. Perhaps
+    // we should check on the accuracy of the matrix quantities that we provide
+    // to DART.
+    if ("dart" == _physicsEngine)
+    {
+      tolerance = 2e-3;
+    }
+
     // Wait for message to be received
-    while (link->WorldForce() != forceWorld ||
-           link->WorldTorque() != torqueWorld)
+    while (!link->WorldForce().Equal(forceWorld, tolerance) ||
+           !link->WorldTorque().Equal(torqueWorld, 1e-6) )
     {
       world->Step(1);
       common::Time::MSleep(1);
     }
 
     // Check force and torque (at CoG?) in world frame
-    EXPECT_EQ(link->WorldForce(), forceWorld);
-    EXPECT_EQ(link->WorldTorque(), torqueWorld);
+    VEC_EXPECT_NEAR(link->WorldForce(), forceWorld, tolerance);
+    VEC_EXPECT_NEAR(link->WorldTorque(), torqueWorld, 1e-6);
 
     // Reset link's physics states
     link->ResetPhysicsStates();